app/backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
